+++ game_patch/graphics/graphics.cpp
@@ -868,7 +868,7 @@ void GraphicsInit()
 
 void GraphicsDrawFpsCounter()
 {
-    if (g_game_config.fps_counter) {
+    if (g_game_config.fps_counter && !rf::is_hud_hidden) {
         auto text = StringFormat("FPS: %.1f", rf::current_fps);
         rf::GrSetColor(0, 255, 0, 255);
         rf::GrDrawAlignedText(rf::GR_ALIGN_RIGHT, rf::GrGetMaxWidth() - 10, 60, text.c_str(), -1, rf::gr_text_material);